sonar tests

+++ src/SFA.DAS.ApprenticeApp.Pwa/Controllers/CmadController.cs
@@ -11,10 +11,12 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
 {
     public class CmadController : Controller
     {
+        private readonly ILogger<AccountController> _logger;
         private readonly IOuterApiClient _client;
 
-        public CmadController(IOuterApiClient client)
+        public CmadController(ILogger<AccountController> logger, IOuterApiClient client)
         {
+            _logger = logger;
             _client = client;
         }
 
@@ -37,6 +39,13 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
                 // Further filter if more than one account is returned               
                 if (apprenticeAccount.Count >= 2) return View("CheckUln");
 
+                var apprentice = apprenticeAccount.SingleOrDefault();
+
+                if (apprentice == null)
+                {
+                    return RedirectToAction("AccountNotFound", "Account");
+                }
+
                 if (model.ApprenticeId == Guid.Empty)
                 {
                     apprenticeDetails = await HandleExistingApprenticeAccount(apprenticeAccount.SingleOrDefault());
@@ -48,7 +57,7 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
 
                 if (apprenticeDetails?.MyApprenticeship != null)
                 {
-                    var viewModel = await ConstructViewModel(apprenticeDetails);
+                    var viewModel = ConstructViewModel(apprenticeDetails);
                     ModelState.Clear();
                     return View("ConfirmApprenticeshipDetails", viewModel);
                 }
@@ -59,6 +68,7 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
             }
             catch (Exception ex)
             {
+                _logger.LogInformation(ex, "Error finding apprentice");
                 return RedirectToAction("AccountNotFound", "Account");
             }
 
@@ -73,7 +83,7 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
 
             var apprenticeDetails = await _client.GetApprenticeDetails(apprenticeship.ApprenticeId);
 
-            var viewModel = await ConstructViewModel(apprenticeDetails);
+            var viewModel = ConstructViewModel(apprenticeDetails);
             ModelState.Clear();            
 
             return View("ConfirmApprenticeshipDetails", viewModel);
@@ -104,6 +114,7 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
             }
             catch (Exception ex)
             {
+                _logger.LogInformation(ex, $"Error confirming apprenticeship details for apprenticeId: {apprenticeId}");
                 return RedirectToAction("CmadError", "Account");
             }
         }
@@ -129,6 +140,12 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
 
             var apprenticeDetails = await _client.GetApprenticeDetails(new Guid(apprenticeAccount.ApprenticeId.ToByteArray()));
 
+            if (apprenticeDetails == null)
+            {
+                _logger.LogInformation($"Apprentice Details not found for {apprenticeId}");
+                return new ApprenticeDetails();
+            }
+
             // Update Provider Apprentice Account with User OneLogin credentials              
             await _client.DeleteApprenticeAccount(existingApprenticeAccount.ApprenticeId);
             await _client.UpdateApprentice(new Guid(apprenticeDetails.Apprentice.ApprenticeId.ToByteArray()), pathDoc);
@@ -139,7 +156,7 @@ namespace SFA.DAS.ApprenticeApp.Pwa.Controllers
             return apprenticeDetails;
         }
 
-        private async Task<ConfirmApprenticeshipDetailsViewModel> ConstructViewModel(ApprenticeDetails apprenticeDetails)
+        private static ConfirmApprenticeshipDetailsViewModel ConstructViewModel(ApprenticeDetails apprenticeDetails)
         {                        
             var model = new ConfirmApprenticeshipDetailsViewModel()
             {
